Cambios en PaintPane: A partir de ahora se utilizan MovableFigure. Cada figura sabe moverse, el frontEnd sólo pide que se lo haga.

+++ src/backend/model/movables/Movable.java
@@ -1,5 +1,5 @@
 package backend.model.movables;
 
 public interface Movable {
-    void move(int moveInX, int moveInY);
+    void move(double moveInX, double moveInY);
 }
